requirements.txt
Django==1.6.2
South==0.8.4
django-bootstrap-form==3.1
django-debug-toolbar==1.0.1
django-extra-views==0.6.4
django-libtech-emailuser==0.2
django-tastypie==0.11.0
itsdangerous==0.23
psycopg2==2.5.2
pyshp==1.2.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.5.2
sqlparse==0.1.11
wsgiref==0.1.2
